requirements-dev.txt
clearml==1.6.3
einops==0.4.1
flake8==4.0.1
isort==5.10.1
jax[cpu]==0.3.15
jaxlib==0.3.15
matplotlib==3.5.2
mypy==0.971
pandas==1.4.3
Pillow==9.2.0
pydocstyle==6.1.1
pylint==2.14.5
pylint-django==2.5.3
pytest-cov==3.0.0
pytorch-ignite==0.4.9
rich==12.5.1
scikit_image==0.19.3
torch==1.8.1
wandb==0.12.21
yaecs==1.0.1
